zaapsign_backend/requirements.txt
Django>=4.0,<5.0
djangorestframework
psycopg2-binary
gunicorn
requests
django-cors-headers==3.13.0  

